Move layers to the new sidebar panel
